version 0.5.1 Fix elevation value type
git-svn-id: https://nmea.svn.sourceforge.net/svnroot/nmea/trunk@10 c96b1883-922c-0410-873a-96fdbb105808

+++ include/nmea/info.h
@@ -94,7 +94,7 @@ typedef struct _nmeaINFO
 
     double  lat;        /**< Latitude in NDEG - +/-[degree][min].[sec/60] */
     double  lon;        /**< Longitude in NDEG - +/-[degree][min].[sec/60] */
-    int     elv;        /**< Antenna altitude above/below mean sea level (geoid) in meters */
+    double  elv;        /**< Antenna altitude above/below mean sea level (geoid) in meters */
     double  speed;      /**< Speed over the ground in kilometers/hour */
     double  direction;  /**< Track angle in degrees True */
     double  declination; /**< Magnetic variation degrees (Easterly var. subtracts from true course) */
